+++ lib/commands/add.php
@@ -11,7 +11,7 @@ use Huxtable\Core\File;
 /**
  * @command		add
  * @desc		Start tracking a new project
- * @usage		add [<group>/]<name> <path>
+ * @usage		add [<group>/]<name> <path> [--url=<url>]
  * @alias		track
  */
 $commandAdd = new CLI\Command( 'add', 'Start tracking a new project', function( $name, $path )
@@ -23,6 +23,34 @@ $commandAdd = new CLI\Command( 'add', 'Start tracking a new project', function( 
 	catch( \Exception $e )
 	{
 		throw new CLI\Command\CommandInvokedException( "Couldn't track project. {$e->getMessage()}", 1 );
+	}
+
+	/* Clone new project to track */
+	$repoURL = $this->getOptionValue( 'url' );
+	if( !is_null( $repoURL ) )
+	{
+		if( $dirProject->exists() )
+		{
+			if( count( $dirProject->children() ) > 0 )
+			{
+				throw new CLI\Command\CommandInvokedException( "Couldn't clone repository: '{$path}' already exists and is not an empty directory.", 1 );
+			}
+		}
+
+		echo "Cloning into '{$dirProject}'... " ;
+		$result = CLI\Shell::exec( "git clone --recursive {$repoURL} {$dirProject}", true, '  > ' );
+
+		if( $result['exitCode'] === 0 )
+		{
+			echo 'done.' . PHP_EOL;
+		}
+		else
+		{
+			echo 'failed:' . PHP_EOL . PHP_EOL;
+			echo $result['output']['formatted'] . PHP_EOL;
+
+			exit( 1 );
+		}
 	}
 
 	if( !$dirProject->exists() )
@@ -46,7 +74,22 @@ $commandAdd = new CLI\Command( 'add', 'Start tracking a new project', function( 
 	return $output->flush();
 });
 
+/* Options */
+$commandAdd->registerOption( 'url' );
+
+/* Aliases */
 $commandAdd->addAlias( 'track' );
-$commandAdd->setUsage( 'add [<group>/]<name> <path>' );
+
+/* Usage */
+$commandAddUsage = <<<USAGE
+add [<group>/]<name> <path> [--url=<url>]
+
+OPTIONS
+    --url=<url>
+        Clone the Git repository at <url> to <path>.
+
+USAGE;
+
+$commandAdd->setUsage( $commandAddUsage );
 
 return $commandAdd;